🚚rename: 디렉토리 구조 전반적으로 변경
