+++ be/src/main/java/com/kiosk/api/product/controller/dto/ProductCategoryResponse.java
@@ -1,10 +1,9 @@
-package com.kiosk.api.product.web.controller.dto;
+package com.kiosk.api.product.controller.dto;
 
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.ToString;
 
 @AllArgsConstructor
