+++ be/src/main/java/com/kiosk/api/product/service/CategoryService.java
@@ -1,4 +1,4 @@
-package com.kiosk.api.product.web.service;
+package com.kiosk.api.product.service;
 
 import com.kiosk.api.product.domain.entity.Category;
 import com.kiosk.api.product.domain.repository.CategoryRepository;
